code refactoring of all utils
