+++ leia-common/src/main/java/com/grookage/leia/common/utils/Utils.java
@@ -2,7 +2,6 @@ package com.grookage.leia.common.utils;
 
 import com.grookage.leia.common.exception.SchemaValidationException;
 import com.grookage.leia.common.exception.ValidationErrorCode;
-import com.grookage.leia.models.annotations.Optional;
 import lombok.experimental.UtilityClass;
 
 import java.lang.reflect.Field;
@@ -25,7 +24,7 @@ public class Utils {
     }
 
     public Type[] getTypeArguments(ParameterizedType parameterizedType) {
-        Type[] typeArguments = parameterizedType.getActualTypeArguments();
+        final var typeArguments = parameterizedType.getActualTypeArguments();
         if (typeArguments.length == 0) {
             throw SchemaValidationException.error(ValidationErrorCode.INVALID_SCHEMAS,
                     String.format("No type arguments found for %s", parameterizedType));
@@ -37,20 +36,5 @@ public class Utils {
         return Arrays.stream(klass.getEnumConstants())
                 .map(enumConstant -> ((Enum<?>) enumConstant).name())
                 .collect(Collectors.toSet());
-    }
-
-    private boolean isOptional(Type type) {
-        if (type instanceof Class<?> klass) {
-            return isOptional(klass);
-        }
-        return false;
-    }
-
-    private boolean isOptional(Class<?> klass) {
-        return klass.isAnnotationPresent(Optional.class);
-    }
-
-    private boolean isOptional(Field field) {
-        return field.isAnnotationPresent(Optional.class);
     }
 }
