+++ leia-common/src/test/java/com/grookage/leia/common/utils/ValidationUtilsTest.java
@@ -1,4 +1,4 @@
-package com.grookage.leia.core.utils;
+package com.grookage.leia.common.utils;
 
 import com.grookage.leia.models.ResourceHelper;
 import com.grookage.leia.models.attributes.ArrayAttribute;
@@ -222,6 +222,7 @@ class ValidationUtilsTest {
                                 .enumclass(Enumclass.ONE)
                         .build()))
                 .build();
+        // TODO::Abhishek Finish this
         final var jsonNode = ResourceHelper.getObjectMapper().valueToTree(testRecord);
     }
