+++ leia-core/pom.xml
@@ -35,6 +35,11 @@
 
         <dependency>
             <groupId>com.grookage.leia</groupId>
+            <artifactId>leia-common</artifactId>
+        </dependency>
+
+        <dependency>
+            <groupId>com.grookage.leia</groupId>
             <artifactId>leia-repository</artifactId>
         </dependency>
