+++ leia-dropwizard/src/main/java/com/grookage/leia/dropwizard/bundle/resources/SchemaResource.java
@@ -19,10 +19,10 @@ package com.grookage.leia.dropwizard.bundle.resources;
 import com.codahale.metrics.annotation.ExceptionMetered;
 import com.codahale.metrics.annotation.Timed;
 import com.fasterxml.jackson.databind.JsonNode;
+import com.grookage.leia.common.utils.ValidationUtils;
 import com.grookage.leia.core.exception.LeiaErrorCode;
 import com.grookage.leia.core.exception.LeiaException;
 import com.grookage.leia.core.retrieval.SchemaRetriever;
-import com.grookage.leia.core.utils.ValidationUtils;
 import com.grookage.leia.models.GenericResponse;
 import com.grookage.leia.models.request.NamespaceRequest;
 import com.grookage.leia.models.schema.SchemaDetails;
