+++ leia-models/src/main/java/com/grookage/leia/models/annotations/attribute/qualifiers/PII.java
@@ -1,4 +1,4 @@
-package com.grookage.leia.models.annotations.qualifiers;
+package com.grookage.leia.models.annotations.attribute.qualifiers;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
@@ -7,6 +7,5 @@ import java.lang.annotation.Target;
 
 @Target({ElementType.TYPE, ElementType.FIELD})
 @Retention(RetentionPolicy.RUNTIME)
-public @interface ShortLived {
-    int ttlSeconds();
+public @interface PII {
 }
